refactor(schedule): extract the claim→fire→heal core into fire-engine (phase 3)
Pulls the exactly-once firing path out of index.ts into lib/fire-engine.ts —
the most cohesive unit in the file and the live fire path for ALL schedules.

Moves MISSED_WINDOW_REASON + ProcessRunResult, getScheduleDefaultOwner,
fireScheduledRun, fireScheduledRunFromPayload, storedTaskExists, SkipReason +
skipFire, runClaimedFire (the gate→create→attach→audit body),
HEAL_AFTER_MS + healPendingCronClaims, and fireManualRun. They all read the
activated ctx through getPluginCtx() (the single cell from phase 2), so no
ctx threading was needed.

index re-imports the five names the scheduler loop, run-now routes/exec tool,
and __scheduleTestInternals still call, and re-exports MISSED_WINDOW_REASON so
the `@bakin/schedule/index` test surface stays stable. It sheds ~12 fire-only
imports (createTaskWithEffects, readTaskboard, the ledger claim/heal verbs,
the sidecar gate verbs, randomUUID, expandTemplate, BridgeResult).

Pure behavior-preserving relocation — functions byte-identical;
ProcessRunResult.body kept as BridgeResult (not loosened to Record). 1,418 -> 1,153.

Verified: typecheck + eslint + the two fire-path tests (cron-dedup +
blocked-fire-routing, 24/0) + full schedule suite (280/0) + full suite
(5124/0) + binary build + DOCKERIZED-RIG E2E against real OpenClaw (isolated
mode): the schedule plugin activates (200 + live jobs — the activation the
env-ENOENT boot smoke can't show); run-now drove the moved path end to end →
success + task created + matching task_created audit row; a second run-now
correctly skipped with reason `overlap` (the moved board-read + skipFire). The
exactly-once claim, gate logic, board IO, and audit all behave against the
live stack.
